Save and access (incident) location data

+++ core/network/src/main/java/com/crisiscleanup/core/network/model/NetworkWorksite.kt
@@ -90,7 +90,7 @@ data class NetworkWorksiteFull(
     @Serializable
     data class Location(
         val type: String,
-        val coordinates: List<Float>,
+        val coordinates: List<Double>,
     )
 
     @Serializable
